nanosaur_robot/requirements.txt
setuptools
pillow
Adafruit-SSD1306
Adafruit_MotorHat